separa as funções da api com forme o modulo da busca
